feat: add AI chat route, implement modular server startup tasks, and introduce automated auction settlement job
